+++ tomcat/src/main/java/org/apache/coyote/http11/Session/Session.java
@@ -16,7 +16,4 @@ public class Session {
         return sessionData.get(sessionId);
     }
 
-    public static void logoutUser(String sessionId) {
-        sessionData.remove(sessionId);
-    }
 }
