--- tomcat/src/main/java/org/apache/coyote/http11/enums/Methods.java
@@ -1,4 +1,0 @@
-package org.apache.coyote.http11.enums;
-
-public class Methods {
-}
